requirements.txt
streamlit
io
Biopython
PyBioMed.Pyprotein
joblib
pandas
sklearn.feature_selection